// frontend/unilocal/src/app/map/interfaces/places.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
    type:        string;
    query:       string[];
    features:    Feature[];
    attribution: string;
}

export interface CustomFeature extends Feature {
    type: "Feature";
}

export interface Feature {
    id: string;
    type: string;
    place_type: string[];
    relevance: number;
    properties: Properties;
    text_es: string;
    language_es?: Language;
    place_name_es: string;
    text: string;
    language?: Language;
    place_name: string;
    bbox?: [number, number, number, number, number, number] | [number, number, number, number];
    center: number[];
    geometry: Geometry;
    context?: Context[];
}

export interface Context {
    id:           string;
    mapbox_id:    string;
    wikidata?:    string;
    short_code?:  string;
    text_es:      string;
    language_es?: Language;
    text:         string;
    language?:    Language;
}

export enum Language {
    Es = "es",
}

export interface Geometry {
    type: "Point" | "MultiPoint" | "LineString" | "MultiLineString" | "Polygon" | "MultiPolygon";
    coordinates: any; // Aquí podrías especificar un tipo más específico según el tipo de geometría
    bbox?: [number, number, number, number, number, number] | [number, number, number, number];
}

export interface Properties {
    mapbox_id?:  string;
    wikidata?:   string;
    foursquare?: string;
    landmark?:   boolean;
    address?:    string;
    category?:   string;
    accuracy?:   string;
    short_code?: string;
}